pyproject.toml
[project]
name = "mcpdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp-use>=0.1.0",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0"
]
